+++ cli/tests/pcluster/test_utils.py
@@ -677,39 +677,39 @@ def test_get_service_principal(mocker, service_name, partition, region, regional
     "partition, service, region, account, resource, expected_result",
     [
         pytest.param(
-            "aws",
-            "s3",
-            "us-east-1",
-            "123456789012",
-            "bucket_name",
-            "arn:aws:s3:us-east-1:123456789012:bucket_name",
+            "ANY_PARTITION",
+            "ANY_SERVICE",
+            "ANY_REGION",
+            "ANY_ACCOUNT_ID",
+            "ANY_RESOURCE",
+            "arn:ANY_PARTITION:ANY_SERVICE:ANY_REGION:ANY_ACCOUNT_ID:ANY_RESOURCE",
             id="All fields provided",
         ),
         pytest.param(
-            "aws",
-            "s3",
+            "ANY_PARTITION",
+            "ANY_SERVICE",
             "",
-            "123456789012",
-            "bucket_name",
-            "arn:aws:s3::123456789012:bucket_name",
+            "ANY_ACCOUNT_ID",
+            "ANY_RESOURCE",
+            "arn:ANY_PARTITION:ANY_SERVICE::ANY_ACCOUNT_ID:ANY_RESOURCE",
             id="Empty region",
         ),
         pytest.param(
-            "aws",
-            "s3",
-            "us-east-1",
+            "ANY_PARTITION",
+            "ANY_SERVICE",
+            "ANY_REGION",
             "",
-            "bucket_name",
-            "arn:aws:s3:us-east-1::bucket_name",
+            "ANY_RESOURCE",
+            "arn:ANY_PARTITION:ANY_SERVICE:ANY_REGION::ANY_RESOURCE",
             id="Empty account",
         ),
         pytest.param(
-            "aws",
-            "s3",
+            "ANY_PARTITION",
+            "ANY_SERVICE",
             "",
             "",
-            "bucket_name",
-            "arn:aws:s3:::bucket_name",
+            "ANY_RESOURCE",
+            "arn:ANY_PARTITION:ANY_SERVICE:::ANY_RESOURCE",
             id="Empty region and account",
         ),
     ],
